Added files via upload
This is a Jupyter notebook with the exploration of webcam image data captured by the scraper. It includes a unique image timeline to check the polling intervals for you data source, an image slider to explore you data and a transient object detection model for counting people in the data snapshot.